backend/requirements.txt
uvicorn
python-multipart
fastapi[all]
ollama
pyyaml
pytest 
httpx 
pytest-mock